requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
PyJWT==2.8.0
pytest==9.0.1
Werkzeug>=2.3.7
